solve no.4 removeFromArray exercise
This one was the most frustrating bcoz of my lack of focus.
What I learned here is the use of forEach method and the !args.includes(item) method.
The crux of the problem was that if the original array's element doesn't exists in the rest parameters.Then simply add main array's items to a new array confirming that they need not be removed.
Thus,the 2 loops work.